Chapter 09/requirements.txt
fastavro~=1.10.0
kafka-python~=2.0.2
Flask~=3.1.0
psycopg2-binary~=2.9.9
pybreaker~=1.2.0